feat: Add a function to check if the instance is the only instance
The function: bool SingleApplication::isSingleInstance()
Returns if the instance is the only running instance.

It is hard to check the number of running instances after a crash,
so this is an alternative solution for itay-grudev#89.

It uses a trick: the shared memory will be released after the last
instance detaches from it.