Use new QNativeIpcKey based QSharedMemory constructor with Qt 6.6 and higher
Switch to the new QNativeIpcKey based QSharedMemory constructor with Qt 6.6 and higher, the old constructor will be deprecated.
This also makes the library work again with the upcoming Qt 6.6 release and higher.
However, there are still issues with releasing the existing memory in cases where the application is forcefully quit or crashed.